% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{heat_misc}
\alias{heat_misc}
\title{Yet another heatmap function}
\usage{
heat_misc(mat, pheno_list = NULL, z_score = TRUE, row_clust = TRUE,
  col_clust = TRUE, rang = c(-3, 3), axis_scale = 1, show_grid = F,
  grid_lty = 3, row_names = T, col_names = T, mar_padding = c(3, 1, -4,
  0), pals = NULL, leg = NULL, ...)
}
\description{
Produces a very clean looking heatmap and allows for lots of customization
}
\examples{
show_plot_chars()
}

