% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{multiplot}
\alias{multiplot}
\title{Multiple plot function}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{A list of ggplot objects}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\description{
Allows one to plot multiple ggplot plot objects on the same plot.
}
\examples{
show_plot_chars()
}

