% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{show_plot_chars}
\alias{show_plot_chars}
\title{Show base plot characters}
\usage{
show_plot_chars()
}
\description{
This function allows you to print lots of plot characters to the screen so
you know what you can choose from
}
\examples{
show_plot_chars()
}

