% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{trim}
\alias{trim}
\title{Trim leading and trailing whitespace}
\usage{
trim(x)
}
\description{
Similar to \emph{trim} functions in other languages, \code{trim} removes
leading and trailing whitespace from character vectors.
}
\examples{
spacey_string = " hi there  "
trim(spacey_string)
}

