% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{xlsx2csv}
\alias{xlsx2csv}
\title{Wrapper to xlsx2csv command line client}
\usage{
xlsx2csv(infile, outfile, delimiter = "tab", sheet_num = NULL)
}
\arguments{
\item{infile}{Path to xlsx file to convert.}

\item{outfile}{Path to output file.}

\item{delimiter}{Field separator to be used in the output file.}

\item{sheet_num}{Sheet number to convert. Users must provide this or}
}
\description{
\link[=https://github.com/dilshod/xlsx2csv]{xlsx2csv} is an excellent
command line utility for converting \emph{Excel} files to CSV or TSV format.
This function provides an easy interface to that utility. Users must have
xlsx2csv installed via \emph{pip}.
}
\examples{
spacey_string = " hi there  "
trim(spacey_string)
}

