% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{heat_misc}
\alias{heat_misc}
\title{Yet another heatmap function}
\usage{
heat_misc(mat, pheno_list = NULL, z_score = TRUE, row_clust = TRUE,
  col_clust = TRUE, rang = c(-3, 3), axis_scale = 1, show_grid = F,
  grid_lty = 3, row_names = T, col_names = T, mar_padding = c(3, 1, -4,
  0), pals = NULL, leg = NULL, cfunc = NULL, ...)
}
\arguments{
\item{mat}{A matrix of values to plot in a heatmap.}

\item{pheno_list}{A named list of annotation attributes to add as annotation
tracks. Similar to \code{gplots::heatmap.2} \code{colSideColors}.}

\item{z_score}{Should the rows of the matrix should be z-scored?}

\item{row_clust}{Sould the rows of the matrix be clustered?}

\item{col_clust}{Sould the columns of the matrix be clustered?}

\item{rang}{The maximum and minimum values to be plotted. Values exceeding
this range will be thresholded to the min/max values provided.}

\item{axis_scale}{Value to pass to \code{cex.axis} to change the size of the
axis text}

\item{show_grid}{Should the heatmap cells be outlined with a grid?}

\item{grid_lty}{Integer to select base graphics line type
for \code{show_grid}}

\item{row_names}{Should the row names of the matrix be used as labels?}

\item{col_names}{Should the column names of the matrix be used as labels?}

\item{mar_padding}{Values to alter default \code{par()$mar} values.
\code{c(bottom, left, top, right)}}

\item{pals}{A vector of valid \pkg{RColorBrewer} pallete names to use as
colors for \code{pheno_list}}

\item{leg}{A logical vector indicating which of the \code{pheno_list}
annotations should have a legend included.}

\item{cfunc}{A \code{colorRampPalette} to color the heatmap with.}
}
\description{
\code{heat_misc} Produces a very clean looking heatmap and allows for lots
of customization. Born out of my frustration with bloated heatmap functions
like \code{NMF::aheatmap} and \code{gplots::heatmap.2} that make the creation
of nice looking heatmaps possible, but do now allow for much customization.
}
\examples{
# a matrix
test = matrix(rnorm(100), ncol = 10)
colnames(test) = rownames(test) = letters[1:10]

# basic usage
heat_misc(test)
heat_misc(test, z_score = FALSE)
heat_misc(test, row_clust = FALSE)

# adding annotation
pl = list('pheno' = rep(c('a', 'b', 'c'), length.out = 10))
heat_misc(test, pheno_list = pl)
heat_misc(test, pheno_list = pl, leg = TRUE)
heat_misc(test, pheno_list = pl, leg = TRUE, show_grid = TRUE)
}

